Maintain solver state while minimizing
The previous approach added assertions that would affect future queries
(in practice, constraining the size of each sort). Instead of making
these permanent, we maintain them only for the duration of the
minimization and pass them to check-sat-assuming as assumptions, so
that the solver state is unaffected by the minimization (other than some
extra indicator variables, which we may want to filter out from the user).

Signed-off-by: Tej Chajed <tchajed@vmware.com>