Fix model parsing bug with auxilliary functions
The model parsing code didn't handle auxilliary definitions that aren't
part of the signature (there was a comment about this, but we didn't
have a test case that exercised it). Now, we don't construct an
interpretation for any symbols in the model that aren't in the signature
(though their raw form is still available for evaluation).

Fixes #5.

Signed-off-by: Tej Chajed <tchajed@vmware.com>