Use a script to download solvers
The script downloads just the binary releases to solvers/ in the source
directory. It handles Linux, macOS, and macOS arm64.

After checking the dedicated environment variable overrides, we now
first look in the solvers subdirectory of the source (by embedding the
CARGO_MANIFEST_DIR environment variable at build time).

The tests make sure that the versions being used match the versions
that the script downloads.

Signed-off-by: Tej Chajed <tchajed@vmware.com>