Print solver (error) response if it exits early
We now give the last error reported by CVC5 if it exits early due to an
error in the SMTLIB input.

CVC5 will exit if it encounters an error (eg, ill-typed or malformed
input). Formerly this would confuse the code looking for a `<<DONE>>`
marker, which returned an uninformative UnexpectedClose error, when in
fact we had a buffer of response so far that should be reported to the
user.

Fixes #15.

Signed-off-by: Tej Chajed <tchajed@vmware.com>