Print solver (error) response if it exits early (#21)
We now give the last error reported by CVC5 if it exits early due to an error in the SMTLIB input. For Z3, instead of a panic due to an unexpected check-sat response we now detect and parse error and return an `Err(SolverError::UnexpectedClose(msg))` (this is a bit of an abuse of that constructor which can be cleaned up later if desired).

CVC5 will exit if it encounters an error (eg, ill-typed or malformed input). Formerly this would confuse the code looking for a `<<DONE>>` marker, which returned an uninformative UnexpectedClose error, when in fact we had a buffer of response so far that should be reported to the user.

Z3 would not exit so we could parse its response, but the code that parses a `check-sat` response wasn't looking for error messages and would simply panic.

Signed-off-by: Tej Chajed <tchajed@vmware.com>