Move CI checks to a script
`./tools/ci-check.sh` does a build, runs the tests, and runs `cargo
fmt`. In CI we run `./tools/ci-check.sh --ci` which is basically the
same but more verbose. (Note that this does make the script itself
implement each check twice, which need to be kept in sync.)

There are a few improvements compared to the commands used earlier, from
more closely inspecting the code. First, we use `cargo build --tests` to
also compile the dev dependencies (this is a separate phase mainly so
the test output is cleaner and easier to read). Second, it turns out
that `cargo test --benches` also runs the unit tests, so we were running
them twice; the correct invocation is `cargo test --tests --benches` to
run unit tests + integration tests + benches.

Signed-off-by: Tej Chajed <tchajed@vmware.com>