Speed up model evaluation by using Interpretations (#38)
Model evaluation could be very slow for Z3 models because it sometimes uses large auxilliary functions, which are evaluated over and over again while enumerating all the inputs for other functions. This is exacerbated by the fact that its models aren't minimized, so the universes are large - see #37 for some notes on that.

In order to speed up evaluation, this PR first constructs an Interpretation for every model symbol (not just the ones in the signature), and then it evaluates calls to previously evaluated functions using their Interpretation rather than by substituting their definition.

I added a new test which is just one auxilliary definition and one additional symbol, extracted from the large model. This test used to take 13s and now takes about 0.1s (both in release mode). Even in a debug build (which is the default for tests) this test takes under a second.

The old test now takes 5s in debug mode and has been un-ignored.

This PR isn't perfect because the model symbols aren't evaluated on demand. In order to get any performance benefit, a symbol has to be interpreted before it's used in other definitions, which requires evaluating in topological order. In order to make this code work at all I added a hack that sorts the auxilliary definitions first (those that aren't in the signature and only in the model), which makes these examples fast and deterministic compared to HashMap iteration order. Ultimately we should evaluate on demand, though, which is much more robust.

In addition to this big change, a few performance improvements to the rest of the evaluation code also made a difference:

- Short circuit evaluating `(and ...)` and `(or ...)`.
- Substitute sexps in-place rather than with a copy.
- Combine substitution with evaluation. This has one of the biggest
  performance impacts - it means we don't have to make a whole traversal
  for substitution and can essentially just do the minimum HashMap
  lookup work once at the leaves while anyway evaluating an expression.
  Combined with short circuiting, probably some substitutions are never
  even performed.

Fixes #26.

Signed-off-by: Tej Chajed <tchajed@vmware.com>