Track and report global timing statistics (#47)
The timing reports look like this:

```
total                 : 103.6s
  non-solver          : 39.1s
  solver total        : 64.5s 1342 calls
    check-sat (unsat) : 28.2s 1132 calls
    check-sat (sat)   : 22.1s  105 calls
    get-minimal-model : 14.2s  105 calls
```

The timing for each call is recorded so it would be easy to add some
distribution info here (for example, maybe the 90th percentile of
timings for check-sat calls).

For now, the focus is on breaking down time spent in the solver. For
convenience we also report total time spent elsewhere to make it easy to
tell if optimization effort should be put into solver usage vs. pure
Rust.

The implementation uses a global static variable in a `Mutex`, purely
for convenience. Eventually we could pass the `Timings` struct around to
eliminate the global variable, but for now this is good enough and still
useful.

Fixes #44.

Signed-off-by: Tej Chajed <tchajed@vmware.com>