Don't expose assignment in Model::eval
Rather than having an Option<Assignment>, split eval into a public
interface with no assignment and an internal one that takes an
assignment.

In order to avoid dealing with `&Assignment`s, I also switched to
`im::HashMap` for assignments. This seems to have no performance impact
on running inference on the `consensus_epr.fly` example (as expected;
the hashmaps involved are extremely small, and the tradeoff between the
two is not large).

Signed-off-by: Tej Chajed <tchajed@vmware.com>