Make Next aware of immutable symbols
This means `p'` can be normalized to `p` if `p` is immutable.
Implementing this required passing the signature to `Next` wherever it
is used.

The tests don't use this all that much (in particular not all the way
from a fly file to the SMT solver), so we should probably add a test of
some sort.

Signed-off-by: Tej Chajed <tchajed@vmware.com>