Report non-solver run time using getrusage (#67)
When running with `--time`, previously we would calculate the
"non-solver" time by taking the total end-to-end time and subtracting
the solver time. This isn't accurate when the solver is called
concurrently.

This PR directly uses the `getrusage` system call to get time spent in
the process, excluding children. This isn't quite the same as real time
(it excludes time spent in other processes for example), but it much
better accounts for concurrency. It's the same thing that the Linux
`time` utility would report. (Actually I wanted to know how `time`
worked so I read the source code and that's how I discovered this
syscall in the first place.)

Signed-off-by: Tej Chajed <tchajed@vmware.com>