Implement trie-based data structures for use in inference
Implement a prefix-tree (Trie) which represents a set of sequences, and allows efficient membership checks for sequences, as well as finding all sequences in the set which are subsequences of a given sequence.

Implement a data strcuture which uses a trie to represent a set of sets of sequences, by saving the sequences of all sets in a trie, and remembering which sequences are contained in each set. This data structure (PowerTrie) also allows finding all sets (of sequences) subsumed by another given set (of sequences), where subsumption is defined in the a CNF sense.

Finally, use the above data structure to implement another data structure for quantified CNF formulas. Such a formula is represented as a quantifier prefix and a body, which is a set of sequences (clauses) of literals. Also implement a weakening procedure, which given a model weakens all formulas in the set (in some minimal way) to satisfy it.

Signed-off-by: edenfrenkel <eden060398@gmail.com>