Add trait defining subsumption and implement general maps utilizing it
Add trait defining linear orders with a subsumption relation defined over them. Define the lifting of such orders to and-like and or-like sequences. Implement generic maps with keys from linear orders with subsumption, which also enable getting all keys subsuming/subsumed by any element. Specifically, implement the lifting of such maps for and-like and or-like lifting.

Signed-off-by: edenfrenkel <eden060398@gmail.com>