Systematize advancing frontier frame in qalpha
Make gradually advancing the frontier an explicit option using the flag --gradual-advance. When using this flag, do not consider lemmas which do not appear in UNSAT-cores, (but do drop them if they have no children in the current frame.) Whenever the flag isn't used, advance the frontier as much as possible.

Signed-off-by: edenfrenkel <eden060398@gmail.com>