Sort inference docs tests and tweaks (#52)
In this PR I try to demonstrate what I had in mind with #51 about
documentation and testing by applying those ideas to the sort inference
module. I also made a few material improvements to the code.

In addition, I implemented a checker for the AST invariant that says
that after the sort inference pass runs, there are no AST nodes with
unannotated sorts. I assert that this function returns true at the end
of sort inference. I think this is a good way to handle these kinds of
situations throughout the flyvy codebase.

---------

Signed-off-by: James Wilcox <wilcoxjay@gmail.com>
Signed-off-by: James R. Wilcox <wilcoxjay@gmail.com>