Draft: Add UPDR algorithm (#69)
Added an initial UPDR implementation based on the one in mypyvy. Originally created
as a project in the course "Automatic verification of systems" taught by
Prof. Sharon Shoham Buchbinder.

The implementation itself takes 11.31 secs to run on `lockserver.fly` as
opposed to the 5.47 secs it takes mypyvy to run updr on `lockserv.pyv`.
Sadly, when attempting more complex problems, such as `consensus.fly` ,
we get a timeout. This implementation will require some more work until
it can be merged IMO.

The PR also contains some helper functions that the mypyvy implentation
used and were added to flyvy:
 - Conversion to CNF.
 - Splitting a CNF term into clauses.
 - Negation that pushes the `-` inward in a term.
 - Removal of `prime` from a term.

---------

Signed-off-by: Alex Fischman <alexdfischman@gmail.com>
Signed-off-by: oralmer <almer1or@gmail.com>
Signed-off-by: Or Almer <almer1or@gmail.com>
Signed-off-by: James R. Wilcox <wilcoxjay@gmail.com>
Co-authored-by: Alex Fischman <alexdfischman@gmail.com>
Co-authored-by: James R. Wilcox <wilcoxjay@gmail.com>