Split into crates (#94)
This PR splits the project up into multiple crates, which in theory will
improve compile times. While this is a large change, most of the changes
that I had to make were mechanical, except for the following:

- I made all `mod`s in all `lib.rs` and `mod.rs` public. This causes
lots of missing documentation warnings. I did this because a) the code
does actually need to be documented, and b) I believe that all modules
in a crate should be public, and the module should decide what is
exposed. While the second point is debatable, I don't think the first
point is, so I don't think that this change should be reverted until we
find a way to get the clippy lints without it.
- I removed all of the existing instances of `pub use` in `mod/lib.rs`.
Since all modules are now public, this just creates an extra path to the
same code, so it's unnecessary.
- I added some `pub use`s to the `lib.rs` in `solver`, so that clients
of `solver` would not have to also depend on `smtlib`.
- We split the code up into crates as follows: `fly` and `smtlib` define
common functionality among other crates; `bounded`, `inference`,
`solver`, and `verify` all implement the functionality that the user
cares about; and `temporal-verifier` defines the command line interface.
While most files are in the crate that corresponds to the folder that
they were in before this commit, a few files (such as `timing.rs` and
`concurrent.rs`) got moved due to being in the wrong place in the
dependency ordering.
- Finally, there was one more folder, `term` that we weren't quite sure
what to do with, as it defines common functionality but also depends on
`fly`. For now, we've left it as a submodule of the `fly` crate.

---------

Signed-off-by: Alex Fischman <alexdfischman@gmail.com>
Signed-off-by: James R. Wilcox <wilcoxjay@gmail.com>
Co-authored-by: Alex Fischman <alexdfischman@gmail.com>
Co-authored-by: James R. Wilcox <wilcoxjay@gmail.com>