Cancel solver only during `check_sat` and `get_model` (#105)
The flow is now that cancelling a solver kills it with a signal if it's
in the middle of a check-sat or get-model, and otherwise simply set a
flag `Status::Stopping`. If the solver is stopping, then on any
subsequent call the solver will be killed. Calls that don't return
anything like `assert` and `push` will silently succeed in this case.
Eventually a `check_sat` or `get_model` will be called, and it will
immediately return `Err(SolverError::Killed)`, and this will be
processed as usual by code that deals with cancellation (if the caller
knows the solver is never cancelled they can treat this, they can panic
on any error to correctly handle I/O errors, which can still happen at
any time).

The result is that all the commands that set up the solver should never
fail (and we can panic if they do), only `check_sat` and `get_model`. It
does mean we no longer handle killing the solver externally with a
signal gracefully - cancellation must go through the API to work.

---------

Signed-off-by: Tej Chajed <tchajed@vmware.com>