Transition system extraction (#110)
Fixes #102.
Adds a new module in `fly` that splits a `Module` up into inits, trs,
axioms, and proofs.
Replaces the five sites mentioned in the issue with the new function, as
well as Houdini and concurrent.rs.
Also, this removes the ability for `verify` to handle non-always
assertions. This is okay because all existing .fly files that didn't
have an always were one-state, and so they could be converted to an
always on immutable relations. However, this does technically mean that
this PR removes functionality of `verify` to handle n-state, non-always
assertions.

---------

Signed-off-by: Alex Fischman <alexdfischman@gmail.com>
Co-authored-by: Alex Fischman <alexdfischman@gmail.com>