Add smart constructors for terms, and other improvements to fly/src/syntax.rs (#131)
Add smart constructors for `Sort`, `Binder`, and all `Term` cases. The
smart constructors can take either something owned, and then they don't
clone, or a reference, and then they clone. This is meant to make code
for programmatically constructing Terms nicer, without a bunch of
`.clone` everywhere. In a few (documented) cases, the smart constructors
do a bit more normalization (e.g., `Term::and([])` returns true and not
an actual `Term::And`).

For now, these smart constructors are only used in code I still haven't
pushed (in the liveness to safety), but I think they should be used
everywhere. Basically, there's no reason to directly construct a `Term`.

To convert existing code, change the constructor to a smart constructor,
and try to replace `arg.clone()` by `&arg` to make calling code look
nicer.

Other small changes:
* Rename `Sort::Id` to `Sort::Uninterpreted`
* Rename `UOp::Previously` to `UOp::Previous`, which I think is more
standard ("previously" is sometimes used for the past version of
eventually)
* Documentation improvements

Minor issues and future TODOs:
* Binders passed to `forall` and `exist` still need to be owned. We can
change this by having `impl From<&Binder> for <Binder>` similarly to
what we do for `Sort` and `Term`.
* `Term::and` and `Term::or` take an `IntoIterator` which forces
uniformity. So you cannot do `Term::and([&t1, t2])` (where you can do
`Term::iff(&t1, t2)`, and must instead do `Term::and([&t1, &t2])` which
introduces an unnecessary clone. I don't think this clone gets optimized
away. If we think this is a concern, we can implement variadic `and` and
`or`, either with specialized functions (say for 2-9 arguments) or
perhaps with a macro.

---------

Signed-off-by: Tej Chajed <tchajed@vmware.com>
Signed-off-by: Oded Padon <oded.padon@gmail.com>
Co-authored-by: Tej Chajed <tchajed@vmware.com>