Use multiple solvers in qalpha (#140)
This defines a `BasicSolver` trait in `solver/src/basics.rs`, which
allows making very basic `check-sat` queries which return a `sat` result
together with a trace, an `unsat` result together with an unsat-core, or
`unknown`. By abstracting this basic functionality it is possible to
implement this trait for a single `SolverConf`, as well as for a vector
of `SolverConf`'s in two different ways, once as `FallbackSolvers` tried
sequentially and second as `ParallelSolvers` tried in parallel. This is
utilized in `qalpha`, which now uses `ParallelSolvers` for its
inductiveness and safety checks by default, and is configurable to use
`FallbackSolvers` instead via `--fallback`. The simulation queries in
`qalpha` still use a single `SolverConf`, now abstracted behind the
`BasicSolver` trait as well.

Solver cancellation has also been moved out of the inference code and
made more generic, using the new `BasicSolverCanceler` trait and
`SolverCancelers` struct, which aggregates multiple
`BasicSolverCanceler`'s and itself implements `BasicSolverCanceler`.
This is used recursively to create hierarchical tree-like cancellation,
where the cancellation of a node cancels all of its descendants. The
idea is that functions should only cancel queries which they (perhaps
recursively) launched and not assume anything about the behavior of
their callers. When the result is bubbled up to those callers, they
should cancel their queries according to their own desired behavior.
This allows more general applicablity which is independent of the
specific goals of any single use-case, (in this case `qalpha`, which
cancels all solvers on the first `sat` result.)

Another change is that now transition queries parallelize over
disjunctive transitions, with proper cancellation if a `sat` response is
encountered.

---------

Signed-off-by: edenfrenkel <eden060398@gmail.com>