Set axiom handling (#151)
This PR uses BDDs to generate initial states and does partial evaluation
of post states to prove mutable axioms. It also completely changes
non-boolean-relation translation to make it possible to do partial
evaluation. I think that it should be merged, but it's currently blocked
on merging `bounded-refactor`.

It's also worth mentioning that it slightly changes the
`consensus_forall` example, because the version of `rets.rs` that it
implements can't handle nested non-boolean-returning relations.

---------

Signed-off-by: Alex Fischman <alexdfischman@gmail.com>