Add basic implementation of contexts and a CHC format
This commit introduces the notion of contexts (from Formal Concept Analysis) for computing the least fixpoint in logical domains of formulas. It also adds a CHC format, and a conversion from an FOModule (a first-order transition system) to a system of CHCs, which the context-based least-fixpoint algorithm attempts to solve.

Signed-off-by: edenfrenkel <eden060398@gmail.com>